+++ src/Comparison/ArrayComparison.php
@@ -47,25 +47,30 @@ class ArrayComparison extends AbstractComparison
     {
         return !$this->intersect($array1, $array2);
     }
-    
+
     /**
      * @param array $policyRuleAttributes
      * @param array $attributes
+     * @param array $extraData
      * @return boolean
      */
     public function contains($policyRuleAttributes, $attributes, $extraData = []) {
         foreach($extraData['attribute']->getValue() as $attribute) {
             $result = true;
+            // For each attribute, we check the whole rules set
             foreach($policyRuleAttributes as $pra) {
                 $attributeData = $pra->getAttribute();
                 $attributeData->setValue(
                     $this->comparisonManager->getAttributeManager()->retrieveAttribute($attributeData, $extraData['user'], $attribute)
                 );
-                if(!$this->comparisonManager->compare($pra)) {
+                // If one field is not matched, the whole attribute is rejected
+                if(!$this->comparisonManager->compare($pra, true)) {
+                    //var_dump($attributeData->getName(), $attributeData->getValue(), $pra->getValue());
                     $result = false;
                     break;
                 }
             }
+            // If the result is still true at the end of the attribute check, the rule is enforced
             if($result === true) {
                 return true;
             }