+++ src/Manager/PolicyRuleManager.php
@@ -24,12 +24,12 @@ class PolicyRuleManager
 
     /**
      * @param string $ruleName
-     *
+     * @param object $user
+     * @param object $resource
      * @return PolicyRule
-     *
      * @throws \InvalidArgumentException
      */
-    public function getRule($ruleName)
+    public function getRule($ruleName, $user, $resource)
     {
         if(!isset($this->rules[$ruleName])) {
             throw new \InvalidArgumentException('The given rule "' . $ruleName . '" is not configured');
@@ -39,18 +39,20 @@ class PolicyRuleManager
             ->setName($ruleName)
         ;
         // For each policy rule attribute, the data is formatted
-        foreach($this->processRuleAttributes($this->rules[$ruleName]['attributes']) as $pra) {
+        foreach($this->processRuleAttributes($this->rules[$ruleName]['attributes'], $user, $resource) as $pra) {
             $rule->addPolicyRuleAttribute($pra);
         }
         return $rule;
     }
-    
+
     /**
      * This method is meant to convert attribute data from array to formatted policy rule attribute
-     * 
+     *
      * @param array $attributes
+     * @param object $user
+     * @param object $resource
      */
-    public function processRuleAttributes($attributes) {
+    public function processRuleAttributes($attributes, $user, $resource) {
         foreach($attributes as $attributeName => $attribute) {
             $pra = (new PolicyRuleAttribute())
                 ->setAttribute($this->attributeManager->getAttribute($attributeName))
@@ -58,6 +60,7 @@ class PolicyRuleManager
                 ->setComparisonType($attribute['comparison_type'])
                 ->setValue((isset($attribute['value'])) ? $attribute['value'] : null)
             ;
+            $this->processRuleAttributeComparisonType($pra, $user, $resource);
             // In the case the user configured more keys than the basic ones
             // it will be stored as extra data
             foreach($attribute as $key => $value) {
@@ -67,6 +70,20 @@ class PolicyRuleManager
             }
             // This generator avoid useless memory consumption instead of returning a whole array
             yield $pra;
+        }
+    }
+
+    /**
+     * This method is meant to set appropriated extra data to $pra depending on comparison type
+     *
+     * @param PolicyRuleAttribute $pra
+     * @param object $user
+     * @param object $resource
+     */
+    public function processRuleAttributeComparisonType(PolicyRuleAttribute $pra, $user, $resource) {
+        switch($pra->getComparisonType()) {
+            case 'user': $pra->setExtraData(['user' => $user]); break;
+            case 'object': $pra->setExtraData(['resource' => $resource]); break;
         }
     }
 }